fix: correct stale imports after rebase onto main's skill→tool renames
After rebasing onto main (which renamed sdk/skill → sdk/tool and
toolkits/resource/tool_* → skill_*), update imports in resource_toolkit,
local_script_executor, and four test files to use the current paths.
